매물관리 상세검색 기능 및 화면 구성

+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/construct/ConstructInfo.java
@@ -30,7 +30,7 @@ public class ConstructInfo extends BaseTimeEntity {
 
     private int useAprDay;
 
-    private int platArea;
+    private double platArea;
 
     private Double archArea;
